--- example_insert_and_work_test.go
@@ -1,103 +1,0 @@
-package river_test
-
-import (
-	"context"
-	"fmt"
-	"log/slog"
-	"sort"
-
-	"github.com/jackc/pgx/v5/pgxpool"
-
-	"github.com/riverqueue/river"
-	"github.com/riverqueue/river/riverdbtest"
-	"github.com/riverqueue/river/riverdriver/riverpgxv5"
-	"github.com/riverqueue/river/rivershared/riversharedtest"
-	"github.com/riverqueue/river/rivershared/util/slogutil"
-	"github.com/riverqueue/river/rivershared/util/testutil"
-)
-
-type SortArgs struct {
-	// Strings is a slice of strings to sort.
-	Strings []string `json:"strings"`
-}
-
-func (SortArgs) Kind() string { return "sort" }
-
-type SortWorker struct {
-	river.WorkerDefaults[SortArgs]
-}
-
-func (w *SortWorker) Work(ctx context.Context, job *river.Job[SortArgs]) error {
-	sort.Strings(job.Args.Strings)
-	fmt.Printf("Sorted strings: %+v\n", job.Args.Strings)
-	return nil
-}
-
-// Example_insertAndWork demonstrates how to register job workers, start a
-// client, and insert a job on it to be worked.
-func Example_insertAndWork() {
-	ctx := context.Background()
-
-	dbPool, err := pgxpool.New(ctx, riversharedtest.TestDatabaseURL())
-	if err != nil {
-		panic(err)
-	}
-	defer dbPool.Close()
-
-	workers := river.NewWorkers()
-	river.AddWorker(workers, &SortWorker{})
-
-	riverClient, err := river.NewClient(riverpgxv5.New(dbPool), &river.Config{
-		Logger: slog.New(&slogutil.SlogMessageOnlyHandler{Level: slog.LevelWarn}),
-		Queues: map[string]river.QueueConfig{
-			river.QueueDefault: {MaxWorkers: 100},
-		},
-		Schema:   riverdbtest.TestSchema(ctx, testutil.PanicTB(), riverpgxv5.New(dbPool), nil), // only necessary for the example test
-		TestOnly: true,                                                                         // suitable only for use in tests; remove for live environments
-		Workers:  workers,
-	})
-	if err != nil {
-		panic(err)
-	}
-
-	// Out of example scope, but used to wait until a job is worked.
-	subscribeChan, subscribeCancel := riverClient.Subscribe(river.EventKindJobCompleted)
-	defer subscribeCancel()
-
-	if err := riverClient.Start(ctx); err != nil {
-		panic(err)
-	}
-
-	// Start a transaction to insert a job. It's also possible to insert a job
-	// outside a transaction, but this usage is recommended to ensure that all
-	// data a job needs to run is available by the time it starts. Because of
-	// snapshot visibility guarantees across transactions, the job will not be
-	// worked until the transaction has committed.
-	tx, err := dbPool.Begin(ctx)
-	if err != nil {
-		panic(err)
-	}
-	defer tx.Rollback(ctx)
-
-	_, err = riverClient.InsertTx(ctx, tx, SortArgs{
-		Strings: []string{
-			"whale", "tiger", "bear",
-		},
-	}, nil)
-	if err != nil {
-		panic(err)
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		panic(err)
-	}
-
-	waitForNJobs(subscribeChan, 1)
-
-	if err := riverClient.Stop(ctx); err != nil {
-		panic(err)
-	}
-
-	// Output:
-	// Sorted strings: [bear tiger whale]
-}
